Fixed php5 parser include
