+++ .gitignore
@@ -172,3 +172,4 @@ ext/phalcon/mvc/view/engine/volt/parser.php5.out
 ext/phalcon/mvc/view/engine/volt/parser.php7.out
 ext/phalcon/mvc/model/query/parser.php5.out
 ext/phalcon/mvc/model/query/parser.php7.out
+ext/phalcon/mvc/model/query/.libs/
