+++ ext/phalcon/mvc/model/query/parser.php7.inc.h
@@ -1,0 +1,424 @@
+
+#ifdef HAVE_CONFIG_H
+#include "config.h"
+#endif
+
+#include "php.h"
+#include "php_phalcon.h"
+#include "phalcon.h"
+
+#include "parser.php7.h"
+#include "scanner.h"
+#include "phql.h"
+
+#include "kernel/main.h"
+#include "kernel/memory.h"
+#include "kernel/fcall.h"
+#include "kernel/exception.h"
+
+#define PHQL_DEFINE_INIT_ZVAL(var) zval __tmp##var ; zval *var = &__tmp##var;
+#define phql_add_assoc_stringl(var, index, str, len, copy) add_assoc_stringl(var, index, str, len);
+
+static zval *phql_ret_literal_zval(int type, phql_parser_token *T)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init_size(ret, 2);
+	add_assoc_long(ret, "type", type);
+	if (T) {
+		phql_add_assoc_stringl(ret, "value", T->token, T->token_len, 0);
+		efree(T);
+	}
+
+	return ret;
+}
+
+static zval *phql_ret_placeholder_zval(int type, phql_parser_token *T)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init_size(ret, 2);
+	add_assoc_long(ret, "type", type);
+	phql_add_assoc_stringl(ret, "value", T->token, T->token_len, 0);
+	efree(T);
+
+	return ret;
+}
+
+static zval *phql_ret_qualified_name(phql_parser_token *A, phql_parser_token *B, phql_parser_token *C)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init_size(ret, 4);
+
+	add_assoc_long(ret, "type", PHQL_T_QUALIFIED);
+
+	if (A != NULL) {
+		phql_add_assoc_stringl(ret, "ns-alias", A->token, A->token_len, 0);
+		efree(A);
+	}
+
+	if (B != NULL) {
+		phql_add_assoc_stringl(ret, "domain", B->token, B->token_len, 0);
+		efree(B);
+	}
+
+	phql_add_assoc_stringl(ret, "name", C->token, C->token_len, 0);
+	efree(C);
+
+	return ret;
+}
+
+static zval *phql_ret_raw_qualified_name(phql_parser_token *A, phql_parser_token *B)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init(ret);
+
+	add_assoc_long(ret, "type", PHQL_T_RAW_QUALIFIED);
+	if (B != NULL) {
+		phql_add_assoc_stringl(ret, "domain", A->token, A->token_len, 0);
+		phql_add_assoc_stringl(ret, "name", B->token, B->token_len, 0);
+		efree(B);
+	} else {
+		phql_add_assoc_stringl(ret, "name", A->token, A->token_len, 0);
+	}
+	efree(A);
+
+	return ret;
+}
+
+static zval *phql_ret_select_statement(zval *S, zval *W, zval *O, zval *G, zval *H, zval *L, zval *F)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init_size(ret, 5);
+
+	add_assoc_long(ret, "type", PHQL_T_SELECT);
+	add_assoc_zval(ret, "select", S);
+
+	if (W != NULL) {
+		add_assoc_zval(ret, "where", W);
+	}
+	if (O != NULL) {
+		add_assoc_zval(ret, "orderBy", O);
+	}
+	if (G != NULL) {
+		add_assoc_zval(ret, "groupBy", G);
+	}
+	if (H != NULL) {
+		add_assoc_zval(ret, "having", H);
+	}
+	if (L != NULL) {
+		add_assoc_zval(ret, "limit", L);
+	}
+	if (F != NULL) {
+		add_assoc_zval(ret, "forUpdate", F);
+	}
+
+	return ret;
+}
+
+static zval *phql_ret_select_clause(zval *distinct, zval *columns, zval *tables, zval *join_list)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init(ret);
+
+	if (distinct) {
+		add_assoc_zval(ret, "distinct", distinct);
+	}
+
+	add_assoc_zval(ret, "columns", columns);
+	add_assoc_zval(ret, "tables", tables);
+
+	if (join_list) {
+		add_assoc_zval(ret, "joins", join_list);
+	}
+
+	return ret;
+}
+
+static zval *phql_ret_distinct_all(int distinct)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+	ZVAL_LONG(ret, distinct);
+
+	return ret;
+}
+
+static zval *phql_ret_distinct(void)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+	ZVAL_TRUE(ret);
+
+	return ret;
+}
+
+static zval *phql_ret_order_item(zval *column, int sort){
+
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init(ret);
+	add_assoc_zval(ret, "column", column);
+	if (sort != 0 ) {
+		add_assoc_long(ret, "sort", sort);
+	}
+
+	return ret;
+}
+
+static zval *phql_ret_limit_clause(zval *L, zval *O)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init_size(ret, 2);
+
+	add_assoc_zval(ret, "number", L);
+
+	if (O != NULL) {
+		add_assoc_zval(ret, "offset", O);
+	}
+
+	return ret;
+}
+
+static zval *phql_ret_for_update_clause()
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+	ZVAL_BOOL(ret, 1);
+
+	return ret;
+}
+
+static zval *phql_ret_insert_statement(zval *Q, zval *F, zval *V)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init(ret);
+
+	add_assoc_long(ret, "type", PHQL_T_INSERT);
+	add_assoc_zval(ret, "qualifiedName", Q);
+	if (F != NULL) {
+		add_assoc_zval(ret, "fields", F);
+	}
+	add_assoc_zval(ret, "values", V);
+
+	return ret;
+}
+
+static zval *phql_ret_update_statement(zval *U, zval *W, zval *L)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init(ret);
+
+	add_assoc_long(ret, "type", PHQL_T_UPDATE);
+	add_assoc_zval(ret, "update", U);
+	if (W != NULL) {
+		add_assoc_zval(ret, "where", W);
+	}
+	if (L != NULL) {
+		add_assoc_zval(ret, "limit", L);
+	}
+
+	return ret;
+}
+
+static zval *phql_ret_update_clause(zval *tables, zval *values)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init_size(ret, 2);
+	add_assoc_zval(ret, "tables", tables);
+	add_assoc_zval(ret, "values", values);
+
+	return ret;
+}
+
+static zval *phql_ret_update_item(zval *column, zval *expr)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init_size(ret, 2);
+	add_assoc_zval(ret, "column", column);
+	add_assoc_zval(ret, "expr", expr);
+
+	return ret;
+}
+
+static zval *phql_ret_delete_statement(zval *D, zval *W, zval *L)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init(ret);
+
+	add_assoc_long(ret, "type", PHQL_T_DELETE);
+	add_assoc_zval(ret, "delete", D);
+	if (W != NULL) {
+		add_assoc_zval(ret, "where", W);
+	}
+	if (L != NULL) {
+		add_assoc_zval(ret, "limit", L);
+	}
+
+	return ret;
+}
+
+static zval *phql_ret_delete_clause(zval *tables)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init_size(ret, 1);
+	add_assoc_zval(ret, "tables", tables);
+
+	return ret;
+}
+
+static zval *phql_ret_zval_list(zval *list_left, zval *right_list)
+{
+    HashTable *list;
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init(ret);
+
+	list = Z_ARRVAL_P(list_left);
+	if (zend_hash_index_exists(list, 0)) {
+#if PHP_VERSION_ID < 70000
+        HashPosition pos;
+
+		zend_hash_internal_pointer_reset_ex(list, &pos);
+		for (;; zend_hash_move_forward_ex(list, &pos)) {
+
+			zval ** item;
+
+			if (zend_hash_get_current_data_ex(list, (void**)&item, &pos) == FAILURE) {
+				break;
+			}
+
+			Z_ADDREF_PP(item);
+			add_next_index_zval(ret, *item);
+
+		}
+		zval_ptr_dtor(&list_left);
+#endif
+	} else {
+		add_next_index_zval(ret, list_left);
+	}
+
+	if (right_list) {
+		add_next_index_zval(ret, right_list);
+	}
+
+	return ret;
+}
+
+static zval *phql_ret_column_item(int type, zval *column, phql_parser_token *identifier_column, phql_parser_token *alias)
+{
+
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init(ret);
+	add_assoc_long(ret, "type", type);
+	if (column) {
+		add_assoc_zval(ret, "column", column);
+	}
+	if (identifier_column) {
+		phql_add_assoc_stringl(ret, "column", identifier_column->token, identifier_column->token_len, 0);
+		efree(identifier_column);
+	}
+	if (alias) {
+		phql_add_assoc_stringl(ret, "alias", alias->token, alias->token_len, 0);
+		efree(alias);
+	}
+
+	return ret;
+}
+
+static zval *phql_ret_assoc_name(zval *qualified_name, phql_parser_token *alias, zval *with)
+{
+
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init(ret);
+	add_assoc_zval(ret, "qualifiedName", qualified_name);
+
+	if (alias) {
+		phql_add_assoc_stringl(ret, "alias", alias->token, alias->token_len, 0);
+		efree(alias);
+	}
+
+	if (with) {
+		add_assoc_zval(ret, "with", with);
+	}
+
+	return ret;
+}
+
+static zval *phql_ret_join_type(int type)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+	ZVAL_LONG(ret, type);
+
+	return ret;
+}
+
+static zval *phql_ret_join_item(zval *type, zval *qualified, zval *alias, zval *conditions)
+{
+
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init(ret);
+	add_assoc_zval(ret, "type", type);
+
+	if (qualified) {
+		add_assoc_zval(ret, "qualified", qualified);
+	}
+
+	if (alias) {
+		add_assoc_zval(ret, "alias", alias);
+	}
+
+	if (conditions) {
+		add_assoc_zval(ret, "conditions", conditions);
+	}
+
+	return ret;
+}
+
+static zval *phql_ret_expr(int type, zval *left, zval *right)
+{
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init(ret);
+	add_assoc_long(ret, "type", type);
+	if (left) {
+		add_assoc_zval(ret, "left", left);
+	}
+	if (right) {
+		add_assoc_zval(ret, "right", right);
+	}
+
+	return ret;
+}
+
+static zval *phql_ret_func_call(phql_parser_token *name, zval *arguments, zval *distinct)
+{
+
+	PHQL_DEFINE_INIT_ZVAL(ret);
+
+	array_init(ret);
+	add_assoc_long(ret, "type", PHQL_T_FCALL);
+	phql_add_assoc_stringl(ret, "name", name->token, name->token_len, 0);
+	efree(name);
+
+	if (arguments) {
+		add_assoc_zval(ret, "arguments", arguments);
+	}
+
+	if (distinct) {
+		add_assoc_zval(ret, "distinct", distinct);
+	}
+
+	return ret;
+}
