+++ ext/phalcon/mvc/model/query/scanner.c
@@ -1,4 +1,4 @@
-/* Generated by re2c 0.13.6 on Thu Apr 28 22:27:29 2016 */
+/* Generated by re2c 0.13.6 on Thu Apr 28 22:45:03 2016 */
 // 1 "scanner.re"
 
 /*
